Fix wrong command argument

+++ build_on_server.sh
@@ -1,9 +1,12 @@
 # Get the id of the previous expmath image so that it can be deleted later on
 line=$(sudo docker images | grep expmath)
-old_id=$(echo $a | cut -d " " -f3)
+old_id=$(echo $line | cut -d " " -f3)
 
 # Build image
 sudo docker build -t expmath .
 
-# Delete old image
-sudo docker image rm ${old_id}
+# Delete old image only if building the new one was successfully
+if [[$? -e 0 ]]
+then
+    sudo docker image rm ${old_id}
+fi
